update constants with vision and drive information
